Overhauled system to use "behaviors" more in desc
To make up for the nearly identical code between the Simple and Hunter
Creatures, developed a system labelled behaviors. Essentially a generic
form of updating that can be reapplied to multiple Entities. Added a
custom Point class with various functions that are actually useful such
as distance and rotation. Notably in terms of behavior, SimpleCreatures
know to avoid HunterCreatures and weigh, based on hunger level and some
chance, whether or not to pursue food. Also moved the searching for
entities function to the World class, where it felt appropriate to
belong. Created an EntityType to easily classify different entities
while avoiding instance.